backend/requirements.txt
Flask>=2.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-jwt-extended>=4.5.0
flask-cors>=4.0.0
Werkzeug>=2.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pytest>=7.0.0
gunicorn>=21.0.0 